+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   - Prevent Bonemealing when leaves are placed by players/decayable/Persistence Value
     - Reason for this is fruit should only really be growing when the leaves are attached to a tree
 - Fix Fruit Leaves Crash
+- Fix Yucca Trees not being growable on sand tagged blocks
+- Add Cracked Sand and Sandy Dirt to Sand Block and item tags
 
 # 1.0.5-Beta
 - Workaround for a crash involving fruit leaves decay (No Drops)